Create Print powers of 2 less than n
Imagine you are a computer science teacher and one of your students, Ben, is learning about loops and control structures. You decide to give him a problem to work on as practice.

The problem is as follows: Ben needs to write a program that takes in an integer n as input and prints out all the powers of 2 that are less than n. For example, if n is 10, the program should print out 1, 2, 4, and 8.

Input Format

For each test case, you will be given n as an integer input from the user.

Constraints

0 <= n <= 2^30
Output Format

Print as mentioned in the description.

Sample Input 0

20
Sample Output 0

1 2 4 8 16