Create Odd Even and Divisibility by 3
Take n as an integer input. After this you will be given n numbers as integer inputs and you have to print each time if the number is Even or Odd. And Print “Divisible by 3” if the number is a multiple of 3 and print “Not Divisible by 3” if the number is not a multiple of 3

Input Format

N as an Integer Input

N numbers as an Integer Input

Constraints

1 <= N <= 1000 
1 <= K <= 1000
Output Format

Print as Described in Problem Statement

Sample Input 0

5
9
11
2
6
15
Sample Output 0

Odd Divisible by 3
Odd Not Divisible by 3
Even Not Divisible by 3
Even Divisible by 3
Odd Divisible by 3