Create Maximum Product of 3 Numbers
Once upon a time, there was a mathematician named Max who loved solving number puzzles. One day, he was given an array of integers and was challenged to find the maximum product of three numbers from the array.

Max eagerly accepted the challenge and began working on it. Help Max to find the maximum product of three numbers.

NOTE:- After answering the question, attempt the related question in the linked resource to improve your understanding of the question . Click here

Input Format

An integer N, which is the size of the array.

N integers, depicting the elements of the array.

Constraints

- 3<=N<=1000

- -1000<=arr[i]<=1000
Output Format

Return the maximum product of three numbers from the array.

Sample Input 0

5
-7 3 -5 2 4
Sample Output 0

140
Explanation 0

Product of -7, -5 and 4 will give 140.