Create Reverse_Array
Sophia was an aspiring programmer who had just started learning about arrays. One day, her mentor gave her a task to create an array of size N and reverse it. Sophia was excited to take up the challenge and started working on it immediately.

She defined an array arr[] of size N and filled it with integers. However, she got stuck when it came to reversing the array.

Can you help Sophia with this task? Write a program to define an array of size N, fill it with integers, and reverse the array.

Input Format

An integer N, which is the size of the array arr[]

N integers each in a new line, depicting the elements of the array arr[]

Constraints

0 <= N <= 1000

0 <= arr[i] <= 1000
Output Format

Return the array in reverse order.

Sample Input 0

5
2
4
8
3
5
Sample Output 0

5
3
8
4
2
Explanation 0

Since the array is of length 5 and it's elements are {2, 4, 8, 3, 5} therefore, output will be :- 5 3 8 4 2 ,with each integer in one line