Create Rotate Right
John is an athlete who has been training for a marathon for months. He uses a fitness app to track his progress, which stores his daily running data in an array. However, due to a bug in the app, the data is stored is not in order. To fix this, John needs to rotate the array to the right by a certain number of steps.

Can you help John by writing a program that rotates the array to the right by k steps, where k is a non-negative integer?

NOTE:- After answering the question, attempt the related question in the linked resource to improve your understanding of the question . Click here

Input Format

An integer number N representing size of array

N integers number representing elements of array.

An integer number k.

Constraints

1 <= N <= 10^4

-10^6 <= array[index] <= 10^6

0 <= k < N
Output Format

Return the array after performing the operation.

Sample Input 0

7
1 2 3 4 5 6 7  
2
Sample Output 0

6 7 1 2 3 4 5
Explanation 0

after 1st rotate 7 1 2 3 4 5 6

after 2nd rotate 6 7 1 2 3 4 5