Create zero brings zero
Given a fixed-length integer array arr, duplicate each occurrence of zero, shifting the remaining elements to the right.

Note that elements beyond the length of the original array are not written. Do the above modifications to the input array in place and do not return anything.

NOTE :- After answering the question, attempt the related question in the linked resource to improve your understanding of this question. Click here

Input Format

The first line contains N, i.e. the size of the array.

The second line contains N space-separated positive integers A[i] denoting elements of the array.

Constraints

1 <= N <= 10^4

0 <= arr[i] <= 9
Output Format

Return the desired array.

Sample Input 0

8
1 0 2 3 0 4 5 0
Sample Output 0

1 0 0 2 3 0 0 4 
Explanation 0

After calling your function, the input array is modified to: [1,0,0,2,3,0,0,4]