Create Team Formation
You are given a positive integer array skill of even length N where skill[i] denotes the skill of the ith player. Divide the players into n/2 teams of size 2 such that the total skill of each team is equal.

The chemistry of a team is equal to the product of the skills of the players on that team.

Return the sum of the chemistry of all the teams, or return -1 if there is no way to divide the players into teams such that the total skill of each team is equal.

Input Format

The first line contains N, i.e. the size of the array.

The second line contains N space-separated positive integers skills[i] denoting elements of the array.

Constraints

2 <= N <= 10^5

N is even.

1 <= skill[i] <= 1000
Output Format

Return the sum of the chemistry of all the teams, or return -1 if there is no way to divide the players into teams such that the total skill of each team is equal.

Sample Input 0

6
3 2 5 1 3 4
Sample Output 0

22
Explanation 0

Divide the players into the following teams: (1, 5), (2, 4), (3, 3), where each team has a total skill of 6. The sum of the chemistry of all the teams is: 1 * 5 + 2 * 4 + 3 * 3 = 5 + 8 + 9 = 22.

Sample Input 1

2
3 4 
Sample Output 1

12
Explanation 1

The two players form a team with a total skill of 7. The chemistry of the team is 3 * 4 = 12.