Create Int with Maximum Freq
Mark is a data analyst who is trying to analyze the customer data of a retail company. One of the tasks he needs to perform is to find the most common digit in the customer IDs. The IDs are represented as an array of single-digit integers from 0-9. Mark needs to find the digit that occurs the most in the array in order to identify patterns in customer behavior.

Help Mark and find the digit form the array that occurs maximum number of times.

Input Format

First line contains integer N representing size of array.

Second line contains N integer representing elements of array.

Constraints

1 <= N <= 10^4

0 <= A[i] <= 9
Output Format

Return Value whose freq is maximum.

Sample Input 0

5
1 1 2 3 4
Sample Output 0

1