Create Good String Checker 2
Given a string str, return true if str is a good string, or false otherwise.

A string str is good if all the characters that appear in str have the same number of occurrences (i.e., the same frequency).

Input Format

String as an input.

Constraints

1 <= str.length <= 10^4

str consists of lowercase English letters.
Output Format

Return true and false.

Sample Input 0

abacbc
Sample Output 0

true
Explanation 0

The characters that appear in s are 'a', 'b', and 'c'. All characters occur 2 times in s.