Create Spiral Matrix Transformation
Emily, a woman, was assigned a project that required her to transform an M × N integer matrix. Specifically, she was instructed to shift all the elements of the matrix by one step in a spiral order, beginning from the top left corner and proceeding in a clockwise direction. To complete the task, Emily was given the input matrix and expected to provide the modified matrix as output.

Input Format

First line contains, m and n depicting the size of first matrix.

m*n Integer values, depicting all the elements of matrix.

Constraints

1 <= m and n <=1000

-1000<=mat[i][j]<=1000

Output Format

print the modified matrix.

Sample Input 0

3 3
17 90 51 
1 33 69 
79 84 30
Sample Output 0

33 17 90 
79 1 51 
84 30 69