Create Print All Substrings
There is a young girl named Sarah who loves to solve puzzles and riddles. Today, she is given a challenge by her teacher to find all the possible substrings that could be formed from a given string.

Sarah is excited to take on the challenge, but soon realizes that it is a difficult task. She begins by listing out all the possible substrings manually, but it takes her a long time, and she is not sure if she has found all the possible combinations.

Can you help Sarah and write a program that print all the possibe substrings.

Input Format

Take String as an input.

Constraints

1 <= N <= 10^3
Where N is the length of the input string.
Output Format

Print the total number of substrings possible, where every substring is printed on a single line and hence the total number of output lines will be equal to the total number of substrings.

Sample Input 0

abc
Sample Output 0

a 
ab 
abc 
b 
bc 
c