Create Desired String
Take a string as input. Print the count of all the substrings that start with ‘A’ and end with ‘A’. Also print the length of the longest such substring in the second line. In the third line, print that longest substring.

If no such substring exists, print -1.

Input Format

Take string as an input.

Constraints

1 <= str.length() <= 10^4
Output Format

Count of all the substrings
Length of the longest substring
Print the longest substring
Sample Input 0

ABADA
Sample Output 0

3
5
ABADA
Explanation 0

there are 3 substring ABA , ABADA and ADA. longest substring is ABADA and its length is 5